Major Refactoring:
Created options for samtools_fastq.sh to allow arrays of subsets to be submitted, meaning a fastq file for ALL reads and a fastq file for SIMPLEX_ONLY reads can be created from one command.

Changed basecalling, fixer, and post_basecalling to account for and implement this change.

To do:
Write tsv_condenser.sh
Create a test env that "does" each step of the command for testing.